Update dependent-dropdown.js
When using together with other, 3rd party selects (i.e. https://github.com/cakebake/yii2-bootstrap-select), I need to trigger the refrest of the select after DepDrop change completes. But the event trigger needs to be after the loading class removal, as otherwise the loading class remains on the select for some reason.